Setup apollo provider as network data source
I added a (gitignored) .env file to store some necessary variables. Had to rename the
app.json file to app.config.js so I could setup the dotenv library to do
what it does best

Created an apolloClient util which exports a function that creates a new
ApolloClient with options using the apollo-boost library

I then wrapped the call to the Main component in App.js in an
ApolloProvider passing it a client from the afore mentioned apolloClient
util.